+++ density.c
@@ -10,7 +10,7 @@
 #include "densitykernel.h"
 #include "treewalk.h"
 #include "mymalloc.h"
-#include "garbage.h"
+#include "slotsmanager.h"
 #include "endrun.h"
 #include "timestep.h"
 #include "system.h"
