+++ fof.c
@@ -20,7 +20,7 @@
 #include "endrun.h"
 #include "treewalk.h"
 #include "system.h"
-#include "garbage.h"
+#include "slotsmanager.h"
 #include "densitykernel.h"
 
 /*! \file fof.c
