+++ global.c
@@ -8,7 +8,7 @@
 #include "timefac.h"
 #include "timestep.h"
 #include "cooling.h"
-#include "garbage.h"
+#include "slotsmanager.h"
 #include "endrun.h"
 
 /* global state of system
