+++ hydra.c
@@ -7,7 +7,7 @@
 
 
 #include "allvars.h"
-#include "garbage.h"
+#include "slotsmanager.h"
 #include "treewalk.h"
 #include "densitykernel.h"
 #include "mymalloc.h"
