+++ slotsmanager.h
